Btc registration renamed package
