+++ btc-registration/src/main/kotlin/com/d3/btc/registration/config/BtcRegistrationConfig.kt
@@ -1,4 +1,4 @@
-package registration.btc.config
+package com.d3.btc.registration.config
 
 import config.IrohaConfig
 import config.IrohaCredentialConfig
